requirements.txt
Django
PyYAML
Pillow
